backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.5.0
email-validator==2.3.0
pytest==8.4.2
httpx==0.28.1
